Separate test groups into individual files
